Introduce ParserEngine and move parsing to it
