+++ src/MGR.CommandLineParser/MGR.CommandLineParser.csproj
@@ -99,6 +99,7 @@
     <Compile Include="Extensions\ConverterAttributeExtensions.cs" />
     <Compile Include="Extensions\ConverterExtensions.cs" />
     <Compile Include="Extensions\EnumerableExtensions.cs" />
+    <Compile Include="Extensions\EnumeratorExtensions.cs" />
     <Compile Include="Extensions\PropertyInfoExtensions.cs" />
     <Compile Include="Extensions\StringExtensions.cs" />
     <Compile Include="Extensions\TypeExtensions.cs" />
@@ -117,6 +118,7 @@
     <Compile Include="Extensibility\DependencyInjection\IDependencyResolverScope.cs" />
     <Compile Include="ParserBuilder.cs" />
     <Compile Include="ParserBuilderOptions.cs" />
+    <Compile Include="ParserEngine.cs" />
     <Compile Include="Properties\JetBrains.Annotations.cs" />
     <Compile Include="Parser.cs" />
     <Compile Include="ParserOptions.cs" />
